+++ HISTORY.rst
@@ -1,6 +1,10 @@
 =======
 History
 =======
+2024.5.8 General enhancements
+    * Updated to new calculation handling, with ~/SEAMM/gaussian.ini controlling access
+      to the installed version of Gaussian on the machine.
+    * Added energy and gradients to results to support general use in e.g. energy scans.
 
 2024.1.19: Switched to new way to run Gaussian, added option to just write input file
     * Switched to using the new way to run executables, which supports containers.